Minor refactor
- Changed RoomNameGui logic to prevent passing the scenario through
  params needlessly
- Changed common_data to constants, and added fade times to there
