--- open_dread_rando/common_data.py
@@ -1,12 +1,0 @@
-# commonly-used data that multiple files will require
-
-ALL_SCENARIOS = [
-    "s010_cave", "s020_magma", 
-    "s030_baselab",
-    "s040_aqua",
-    "s050_forest",
-    "s060_quarantine",
-    "s070_basesanc",
-    "s080_shipyard",
-    "s090_skybase"
-]
