+++ open_dread_rando/map_icons.py
@@ -5,7 +5,7 @@ from typing import Tuple, Union, TYPE_CHECKING
 
 from mercury_engine_data_structures.formats import Bmmdef
 
-from open_dread_rando.common_data import ALL_SCENARIOS
+from open_dread_rando.constants import ALL_SCENARIOS
 from open_dread_rando.text_patches import patch_text
 if TYPE_CHECKING:
     from open_dread_rando.patcher_editor import PatcherEditor
